% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parserd.R
\name{write_anki_df}
\alias{write_anki_df}
\title{Write an anki csv}
\usage{
write_anki_df(pkg, path)
}
\arguments{
\item{pkg}{The name of the package whose functions you would like to
memorize}

\item{path}{Path for the exported csv}
}
\value{
Nothing, called for its side effects.
}
\description{
This writes an anki csv which can then be imported into the anki app
}
\examples{
write_anki_df("purrr", "purrr.csv")
}
